refactor: eorzea_time
