+++ lib/ff14_bu_bot.rb
@@ -1,0 +1,7 @@
+lib = File.expand_path("..", __FILE__)
+$LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
+
+require "ruboty/actions/eriones"
+require "ruboty/actions/eorzea_time"
+require "ruboty/handlers/eriones"
+require "ruboty/handlers/eorzea_time"
