--- lib/ruboty/eriones/version.rb
@@ -1,5 +1,0 @@
-module Ruboty
-  module Eriones
-    VERSION = "0.0.1".freeze
-  end
-end
